requirements.txt
selenium
pyyaml
pytz
icalendar